Add incremental1 and incremental2 directories
This adds the files for creating the initial Particle and
SimulationParameters classes in the incremental1 directory and the
updated Array3D and StaggeredGrid classes and tests in the
incremental2 directory. Both directories also include the code for
reading initial particle data from input files.

Notably, the incremental2 directory now has a working
implementation and detailed test of a lot of the particle-to-grid
transfer code for the FLIP/PIC algorithm.