Verify SimulationParameters is never copied
Return value optimization is typically applied by C++ compilers,
but due to details of exactly when it is applied, we could
accidentally copy an object of type SimulationParameters
unnecessarily when returning it from a function and assigning it to
a variable. Prior to C++17, we must declare the copy constructor
in SimulationParameters if we are returning an instance of it and
assigning the return value to a variable even if return value
optimization is being applied. So, we include an assertion failure
in the body of the copy constructor of SimulationParameters to
alert us if that copy constructor is ever called by crashing our
program.