Make Array3D copy constructor private
The Array3D copy constructor was originally made public to allow
functions to return Array3D objects. However, as I decided not to
create and destroy Array3D objects repeatedly during the
simulation, we will keep a few Array3D objects in memory throughout
the simulation and simply update their values rather than creating
and destroying the Array3D objects repeatedly. Since return value
optimization only applies when creating and assigning a new object
to a variable, we will never have a use case for return value
optimization for the Array3D class and thus, there is no need for
the copy constructor in the Array3D class, which would only exist
to make the compiler happy in case we ever return an Array3D object
from a function.