Add particle viewer and movie maker
A fluid simulation viewer is added. Scripts to facilitate the
entire workflow going from the source code to generating an MP4
movie of the resulting simulation to cleaning up files is provided
and descripted in the readme file, HOW_TO_RUN.

The stb_image_write library is included to allow writing of the
OpenGL-generated particle scenes to .png files.